sample-open-clip/pyproject.toml
[project]
name = "sample-open-clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "open-clip-torch>=2.32.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "streamlit>=1.45.1",
    "torchvision>=0.22.1",
]
